+++ src/Agents.Net.Benchmarks/FileManipulation/FileManipulationModule.cs
@@ -25,7 +25,6 @@ namespace Agents.Net.Benchmarks.FileManipulation
         {
             builder.RegisterInstance(terminateAction);
             builder.RegisterType<MessageBoard>().As<IMessageBoard>().InstancePerLifetimeScope();
-            builder.RegisterType<FileCompletedAggregator>().As<Agent>().InstancePerLifetimeScope();
             builder.RegisterType<FileFinder>().As<Agent>().InstancePerLifetimeScope();
             builder.RegisterType<FileManipulator>().As<Agent>().InstancePerLifetimeScope();
             builder.RegisterType<GenericFileCompleter>().As<Agent>().InstancePerLifetimeScope();